+++ src/api_insee/conf.py
@@ -1,0 +1,7 @@
+
+API_VERSION = {
+    "url" : 'https://api.insee.fr',
+
+    "path_token" : '/token',
+    "path_siren" : '/entreprises/sirene/V3/siren'
+}