+++ src/api_insee/criteria/field_exact.py
@@ -1,0 +1,7 @@
+from .field import Field
+
+class FieldExact(Field):
+
+    @property
+    def representation(self):
+        return '%s:"%s"' % (self.name, str(self.value))