+++ src/api_insee/criteria/periodic_field.py
@@ -1,0 +1,8 @@
+from .field import Field
+
+class PeriodicField(Field):
+
+    @property
+    def representation(self):
+        rp = super(PeriodicField, self).representation
+        return 'periode(%s)' % rp