+++ src/api_insee/criteria/raw.py
@@ -1,0 +1,12 @@
+from .base import Base
+
+class Raw(Base):
+
+    def __init__(self, value):
+        self.value = value
+
+    def validate(self):
+        return True
+
+    def toURLParams(self):
+        return self.value