+++ tests/conftest.py
@@ -9,3 +9,6 @@
 """
 
 # import pytest
+SIRENE_API_CONSUMER_KEY    = 'DFUozAz4vYtwhEFfbSmx8mvavzsa'
+SIRENE_API_CONSUMER_SECRET = 'nB_0Cv7nzCs3_FeEtr0D_jObttMa'
+SIRENE_API_TOKEN           = 'b42db1ef-4b19-3818-a6e0-82f7eed776fe'